requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
mysql-connector-python==8.0.33